fixing race condition having to do with atomicity of output file and ack.
There is still a race condition which is completely impossible to avoid AFAIK but it will be much harder to trigger it now.
When it is triggered, and then the job gets re-executed, the result will be two provenance records for the same job,
one of which refers to the partially-completed job. The completed output file will be present in this case so can be recovered
without re-doing processing.

also put a filesystem space check and hardcoded it for now to start failing when there is <100MB free space on the disk,
so as to prevent it from becoming 100% full.