Porting blob extraction to V2. currently it's a command-line script and will need to be Celeryized which should be quite simple.
The deposit service client-side had no dependencies on V1 code so I ported that without modification.

On the server side the deposit service still depends on V1 resolution code and will need to be
substantially rewritten to use the resolver, which will be an opportunity to improve the service
so we can run one deposit service for all time series.